requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
typer==0.12.5
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
pdfminer.six==20240706
markitdown[all]==0.1.3
markdown-it-py==3.0.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0
openai==1.52.0
anthropic==0.37.1
google-generativeai==0.8.2
ollama==0.3.3
rapidfuzz==3.9.7
python-dotenv==1.0.1
Jinja2==3.1.4
rich==13.9.2
praisonaiagents
